Move non test-* files out of test folder
