+++ extensions/amp-viewer-integration/0.1/test/integration/test-amp-viewer-integration.js
@@ -20,7 +20,7 @@ import {
   WindowPortEmulator,
   parseMessage,
 } from '../../messaging/messaging';
-import {ViewerForTesting} from '../viewer-for-testing';
+import {ViewerForTesting} from '../../viewer-for-testing';
 import {getSourceUrl} from '../../../../../src/url';
