+++ extensions/amp-viewer-integration/0.1/test/integration/test-amp-webview-viewer-integration.js
@@ -15,7 +15,7 @@
  */
 
 import {AmpViewerIntegration} from '../../amp-viewer-integration';
-import {WebviewViewerForTesting} from '../webview-viewer-for-testing.js';
+import {WebviewViewerForTesting} from '../../webview-viewer-for-testing.js';
 
 
 describes.sandboxed('AmpWebviewViewerIntegration', {}, () => {
