+++ extensions/amp-viewer-integration/0.1/test/integration/test-viewer-initiated-handshake.js
@@ -16,7 +16,7 @@
 
 import {
   WebviewViewerForTesting,
-} from '../viewer-initiated-handshake-viewer-for-testing';
+} from '../../viewer-initiated-handshake-viewer-for-testing';
 
 
 describes.sandboxed('AmpWebviewViewerIntegration', {}, () => {
